Subscription API Changes based on deep-dive
Changes based on our discussions, fixes required for unambiguous code-gen, adding credential types for outbound push-delivery

- Proper operationIds
- New subscription object attributes
- Rearranged protocol settings such that they inherit from a common base
- Introduced the "basic" filters as distinct filter types
- added "all", "any", "not" filters

I've done some validation of the spec with the Swagger code generator and NSwag. For most languages, the output for the polymorphic constructs like filters and protocol settings and credentials is as expected, meaning that the protocol settings and credentials form a class hierarchy and the generated filter objects can form a graph.